Major alterations and additions
Altered the slides so that they’re actually squares, rather than the
“index card” style they were originally in, the square is now actually
a square. Added a portrait of myself. Also, altered the blacklist so
that internet explorer is the only absolutely blacklisted item. Removed
iOS devices and android so I could test how they actually work with the
site. I’ll re-black list as needed.