added trident to the parse of useragent
Trident, as far as I can tell is something Internet Explorer exclusive.
Considering that:
A.) IE11 tries so hard to obscure itself.
B.) There's no version of IE that can support impress.js

I'm going to include trident on the list. Normally I wouldn't think to
do something so far reaching, but in this case I feel it makes sense.